dbug

+++ server/CampusMapApi/Controllers/TestQuery.cs
@@ -35,7 +35,7 @@ namespace CampusMapApi {
 
           string building = record["building"].As<string>();
           string roomNumber = record["roomNumber"].As<string>();
-          int id = record["id"];
+          int id = (int)record["id"];
           string formattedRoom = $"{building} Room {roomNumber}";
 
           LocationNode node = new LocationNode();
